working

+++ src/main/java/com/example/atlas/comments/CommentsService.java
@@ -2,6 +2,7 @@ package com.example.atlas.comments;
 
 import com.example.atlas.comments.dto.CommentsRequest;
 import com.example.atlas.comments.dto.CommentsResponse;
+import com.example.atlas.exception.ForbiddenException;
 import com.example.atlas.task.Task;
 import com.example.atlas.task.TaskRepo;
 import com.example.atlas.exception.NotFoundException;
@@ -9,7 +10,6 @@ import com.example.atlas.users.UserRepo;
 import com.example.atlas.users.Users;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import org.jspecify.annotations.NonNull;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -22,16 +22,16 @@ public class CommentsService {
     private final CommentsRepo commentsRepo;
     private final UserRepo userRepo;
 
-    public List<CommentsResponse> getComment(Long id) {
+    public List<CommentsResponse> getComments(Long id) {
         return commentsRepo.findCommentsByTask_Id(id).stream()
                 .map((CommentsResponse::new))
                 .toList();
 
     }
 
-    public CommentsResponse createComment(Long id, @NonNull CommentsRequest commentsRequest) {
+    public CommentsResponse createComment(Long id, CommentsRequest commentsRequest,String email) {
         Task task = taskRepo.findById(id).orElseThrow(() -> new NotFoundException("Task not found: " + id));
-        Users users = userRepo.findUsersByName(commentsRequest.getUsername()).orElseThrow(() -> new NotFoundException("User not found: " + commentsRequest.getUsername()));
+        Users users = userRepo.findUsersByEmail(email).orElseThrow(() -> new NotFoundException("User not found with email: " + email));
         Comments comments = new Comments();
         comments.setContent(commentsRequest.getContent());
         comments.setTask(task);
@@ -45,15 +45,21 @@ public class CommentsService {
         return new CommentsResponse(comments);
     }
 
-    public CommentsResponse updateComment(Long id, CommentsRequest request) {
+    public CommentsResponse updateComment(Long id, CommentsRequest request,String email) {
         Comments comments = commentsRepo.findById(id).orElseThrow(() -> new NotFoundException("Comment not found: " + id));
+        if(!comments.getUsers().getEmail().equals(email)){
+            throw new ForbiddenException("It's not your comment");
+        }
         comments.setContent(request.getContent());
         commentsRepo.save(comments);
         return new CommentsResponse(comments);
     }
 
-    public void deleteComment(Long id) {
+    public void deleteComment(Long id,String email) {
         Comments comments = commentsRepo.findById(id).orElseThrow(() -> new NotFoundException("Comment not found: " + id));
+        if(!comments.getUsers().getEmail().equals(email)){
+            throw new ForbiddenException("It's not your comment");
+        }
         commentsRepo.delete(comments);
     }
